added photo upload and password field when try to delete any info.

+++ controllers/directorController.js
@@ -3,7 +3,29 @@ const Movie = require('../models/movie');
 const Director = require('../models/director');
 const Genre = require('../models/genre');
 const MovieInstance = require('../models/movieinstance');
+const Admin = require('../models/admin');
 const asyncHandler = require('express-async-handler');
+const fs = require('fs');
+const path = require('path');
+const multer = require("multer"); // For uploading images
+
+/// ADD IMAGES ///
+
+// Set up multer storage and file name
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, "public/uploads/director");
+  },
+  filename: function (req, file, cb) {
+    cb(null, Date.now() + "-" + file.originalname);
+  },
+});
+
+// Create multer upload instance and add file size limit.
+const upload = multer({ 
+  storage: storage,
+  limits: { fileSize: 300 * 1024 } // 300KB limit
+});
 
 exports.director_list = asyncHandler(async (req,res,next)=>{
     const allDirectors = await Director.find({})
@@ -36,6 +58,7 @@ exports.director_create_get = asyncHandler(async (req,res,next)=>{
 })
 
 exports.director_create_post = [
+  upload.single("image"),
     body("first_name")
   .trim()
   .isLength({min:1})
@@ -67,6 +90,7 @@ exports.director_create_post = [
       last_name: req.body.last_name,
       date_of_birth: req.body.date_of_birth,
       date_of_death: req.body.date_of_death,
+      image: req.file ? req.file.filename : null,
     });
 
     if(!errors.isEmpty()){
@@ -100,15 +124,39 @@ exports.director_delete_get = asyncHandler(async (req, res, next) => {
     });
   });
 
-  exports.director_delete_post = asyncHandler(async (req, res, next) => {
-      await Director.findByIdAndDelete(req.body.directorid);
-      res.redirect("/catalog/directors");
-  });
+  exports.director_delete_post = [
+    body("password", "wrong password"),
 
+    asyncHandler(async (req, res, next) => {
+    const [director] = await Promise.all([
+      Director.findById(req.params.id).exec(),
+    ]);
+    const passwordExists = await Admin.findOne({
+      password:req.body.password
+    }).exec();
+    if(!passwordExists){
+      res.send('wrong password');
+    } else{
+      const imageFileName = director.image;
+      await Director.findByIdAndDelete(req.body.directorid);
+      if(imageFileName){
+        const imagePath = path.join(__dirname, "../public/uploads/director", imageFileName);
+        try {
+          fs.unlinkSync(imagePath);
+      } catch (err) {
+          console.error(`Error deleting image file: ${err.message}`);
+      }
+      }
+      res.redirect("/catalog/directors");
+    }
+  })];
+
+  let director_to_update;
 
   exports.director_update_get = asyncHandler(async (req, res, next) => {
     const director = await Director.findById(req.params.id).exec();
-    
+    director_to_update = director.image;
+
     if (director === null) {
       const err = new Error("Director not found");
       err.status = 404;
@@ -122,6 +170,7 @@ exports.director_delete_get = asyncHandler(async (req, res, next) => {
   });
 
   exports.director_update_post = [
+    upload.single("image"),
     body("first_name")
       .trim()
       .isLength({ min: 1 })
@@ -151,6 +200,7 @@ exports.director_delete_get = asyncHandler(async (req, res, next) => {
         last_name: req.body.last_name,
         date_of_birth: req.body.date_of_birth,
         date_of_death: req.body.date_of_death,
+        image: req.file ? req.file.filename : null,
         _id: req.params.id,
       });
   
@@ -163,6 +213,14 @@ exports.director_delete_get = asyncHandler(async (req, res, next) => {
         return;
       } else {
         await Director.findByIdAndUpdate(req.params.id, director);
+        if(director_to_update){
+          const imagePath = path.join(__dirname, "../public/uploads/director", director_to_update);
+          try {
+            fs.unlinkSync(imagePath);
+        } catch (err) {
+            console.error(`Error deleting image file: ${err.message}`);
+        }
+        }
         res.redirect(director.director_url);
       }
     }),
